refactor(builder): extrahiere module_details.rs aus single/mod.rs
- build_module_details_from_normalized, normalized_module_score,
  normalized_module_grade → single/module_details.rs (~940 Zeilen)
- mod.rs schrumpft auf ~1.992 Zeilen
